users and admins tables

+++ admins/models.py
@@ -3,7 +3,7 @@ import json
 
 class Admins(models.Model):
     id = models.AutoField(primary_key=True)
-    is_del = models.IntegerField(default=0)
+    is_del = models.SmallIntegerField(default=0)
     username = models.CharField(max_length=255)
     phone = models.CharField(max_length=255)
     email = models.CharField(max_length=255)
